refactor: rename provider dir AnimePulse -> MyAnimePulse
Renames src/_provider/AnimePulse -> MyAnimePulse and updates all 7 path imports plus the class aliases (AnimePulseSingle/List, animePulseSearch -> MyAnimePulse*), matching the MyAnimePulse brand used everywhere else. Internal keys left intentionally: the animepulseToken setting and the webpack 'animepulse' page-url key are renaming-neutral (changing the setting key would invalidate stored tokens and force re-auth). tsc clean, webextension builds.
