+++ src/_provider/MyAnimePulse/helper.ts
