+++ src/_provider/MyAnimePulse/list.ts
