+++ src/_provider/MyAnimePulse/search.ts
