+++ src/_provider/MyAnimePulse/single.ts
