+++ src/_provider/singleFactory.ts
@@ -9,7 +9,7 @@ import { Single as KitsuSingle } from './Kitsu/single';
 import { Single as MangaBakaSingle } from './MangaBaka/single';
 import { Single as SimklSingle } from './Simkl/single';
 import { Single as ShikiSingle } from './Shikimori/single';
-import { Single as AnimePulseSingle } from './AnimePulse/single';
+import { Single as MyAnimePulseSingle } from './MyAnimePulse/single';
 import { Single as LocalSingle } from './Local/single';
 
 export function getSingle(url: string) {
@@ -45,7 +45,7 @@ export function getSingle(url: string) {
     return new ShikiSingle(url);
   }
   if (syncMode === 'MYANIMEPULSE') {
-    return new AnimePulseSingle(url);
+    return new MyAnimePulseSingle(url);
   }
   throw 'Unknown sync mode';
 }
