+++ src/index-webextension/myanimepulseOauth.ts
@@ -1,4 +1,4 @@
-import { myAnimePulseOauth } from '../_provider/AnimePulse/oauth';
+import { myAnimePulseOauth } from '../_provider/MyAnimePulse/oauth';
 
 api.settings.init().then(() => {
   myAnimePulseOauth();
